fix: :bug: rename own csv folder to avoid clash with built-in csv
