+++ sprout/app/views/projects_id_metadata_id/step_columns.py
@@ -8,7 +8,7 @@ from django.shortcuts import get_object_or_404, redirect, render
 from sprout.app.forms import ColumnsForm
 from sprout.app.models import Columns, Files, Tables
 from sprout.app.views.projects_id_metadata_id.helpers import create_stepper_url
-from sprout.core.csv.csv_reader import read_csv_file
+from sprout.core.csv_utils.csv_reader import read_csv_file
 
 
 def step_columns(request: HttpRequest, table_id: int) -> HttpResponse:
