+++ sprout/app/views/projects_id_metadata_id/step_file_upload.py
@@ -5,7 +5,7 @@ from django.shortcuts import redirect, render
 
 from sprout.app.models import Columns, Files, Tables
 from sprout.app.views.projects_id_metadata_id.helpers import create_stepper_url
-from sprout.core.csv.csv_reader import read_csv_file
+from sprout.core.csv_utils.csv_reader import read_csv_file
 
 
 def step_file_upload(
