+++ tests/core/test_csv_validator.py
@@ -6,7 +6,7 @@ from unittest import TestCase
 
 import polars as pl
 
-from sprout.core.csv.csv_validator import validate_csv
+from sprout.core.csv_utils.csv_validator import validate_csv
 
 
 class CsvValidatorTests(TestCase):
